Preliminary refactor of updateChildPurposeDescriptions, placing it under the SchATransaction model
